+++ src/khaos/serialization/__init__.py
@@ -5,11 +5,19 @@ from khaos.serialization.avro import (
 )
 from khaos.serialization.base import Serializer
 from khaos.serialization.json import JsonSerializer
+from khaos.serialization.protobuf import (
+    ProtobufSerializer,
+    ProtobufSerializerNoRegistry,
+    field_schemas_to_protobuf,
+)
 
 __all__ = [
     "AvroSerializer",
     "AvroSerializerNoRegistry",
     "JsonSerializer",
+    "ProtobufSerializer",
+    "ProtobufSerializerNoRegistry",
     "Serializer",
     "field_schemas_to_avro",
+    "field_schemas_to_protobuf",
 ]